mod: using time as blog id

+++ postgresql/tables.sql
@@ -1,7 +1,7 @@
 CREATE TYPE file_type AS ENUM ('image', 'audio', 'video', 'text', 'archive', 'application');
 
 CREATE TABLE file (
-  id    BIGSERIAL PRIMARY KEY,
+  id    BIGINT PRIMARY KEY,
   name  VARCHAR(250),
   type  file_type,
   mime  VARCHAR(40),
@@ -14,7 +14,7 @@ CREATE TABLE file (
 CREATE TYPE blog_type AS ENUM ('user', 'group', 'chat');
 
 CREATE TABLE blog (
-  id     SERIAL PRIMARY KEY,
+  id     BIGINT PRIMARY KEY,
   domain VARCHAR(24) UNIQUE,
   name   VARCHAR(128),
   type   blog_type NOT NULL,
@@ -26,9 +26,9 @@ CREATE TABLE blog (
 CREATE TYPE relation_type AS ENUM ('follow', 'manage', 'deny');
 
 CREATE TABLE relation (
-  "from" INT REFERENCES blog (id)
+  "from" BIGINT REFERENCES blog (id)
   ON DELETE CASCADE ON UPDATE CASCADE,
-  "to"   INT REFERENCES blog (id)
+  "to"   BIGINT REFERENCES blog (id)
   ON DELETE CASCADE ON UPDATE CASCADE,
   type   relation_type NOT NULL DEFAULT 'follow',
   UNIQUE ("from", "to")
@@ -39,9 +39,9 @@ CREATE TYPE message_type AS ENUM ('dialog', 'chat', 'wall', 'comment', 'child');
 CREATE TABLE "message" (
   id     BIGINT PRIMARY KEY,
   "type" message_type             NOT NULL,
-  "from" INT REFERENCES blog (id)
+  "from" BIGINT REFERENCES blog (id)
   ON DELETE CASCADE ON UPDATE CASCADE NOT NULL,
-  "to"   INT REFERENCES blog (id)
+  "to"   BIGINT REFERENCES blog (id)
   ON DELETE CASCADE ON UPDATE CASCADE,
   parent BIGINT,
   "text" VARCHAR(8000)            NOT NULL
@@ -59,7 +59,7 @@ CREATE TABLE attachment (
 CREATE TYPE attitude_type AS ENUM ('like', 'hate');
 
 CREATE TABLE attitude (
-  "from"    INT REFERENCES blog (id)
+  "from"    BIGINT REFERENCES blog (id)
   ON DELETE CASCADE ON UPDATE CASCADE NOT NULL,
   "message" BIGINT REFERENCES message (id)
   ON DELETE CASCADE ON UPDATE CASCADE NOT NULL,
